After Deadline Editing
Noticed a mistake the day before deadline and out of curiosity, I have
been trying to solve it and managed to figure it out to day.
Problem-
There was an organsiation error wih the media queries and all I had to
was move the

.child:nth-child(4) {
margin-right: 0;
}

attribute to the 900px media query section instead in oder to properly
centre everything off.